Introduce the http::AuthenticatedClient
The `http::AuthenticatedClient` uses and `oauth::Client` to retrieve
bearer tokens and knows to refresh the token when it gets a `401
Unauthorized` response. This is possible thanks to the introduction of
`oauth::Client::refresh_access_token`. Also, `oauth::Client` is now a
trait, so that we can mock it.

This allows us to handle token expirations correctly even when the token
doesn't come with an `expires_in` field.